refactor(ai): colocate full_ai and economy planners under planning/ (Refs #2289)
Move implementations from lib/src/ shims into planning/ so pipeline
layout matches SPEC/ai/ai-architecture.md without changing public API.
strategic_ai now imports economy_planner from the same directory.
